+++ CHANGELOG.md
@@ -12,6 +12,8 @@
   directly from your web server or when you want full
   control over the upload process.
 - Fix an error in the "options list" command.
+- Fix an error that occurred when running "direct_deploy"
+  with no deployers enabled.
 
 ## 9.1.0 (2025-07-17)