+++ CHANGELOG.md
@@ -38,6 +38,8 @@
   Previously, CLI commands had no effect on these options.
 - Support blob_value options in "options get". Previously,
   these always showed a meaningless "1" for the value.
+- Show line count of BLOB values in "options list" instead
+  of always showing "1".
 
 ## 9.1.0 (2025-07-17)